requirements.txt
# 三大法人監測系統 - Python 套件需求

# 網路請求
requests==2.31.0

# 環境變數管理
python-dotenv==1.0.0

# 排程 (可選，用於本地執行)
schedule==1.2.0

# 時區處理
pytz==2024.1
